revert(adapters): make `AdapterL1.getFullRequiredDepositFee` work with overrides
This reverts commit 9feb1e136335f31b7eda3b2e9ef664b1e29ad366.

BREAKING CHANGE: Remove support from `AdapterL1.getFullRequiredDepositFee` for
considering `overrides.from` as the initiator of the operation. This functionality
was previously used to calculate the full deposit fee for accounts whose private
key is unknown. However, this feature is no longer necessary because
`L1VoidSigner.getFullRequiredDepositFee` is specifically designed to handle such
cases.